+++ libtransmission/metainfo.c
@@ -4,7 +4,6 @@
  * It may be used under the GNU GPL versions 2 or 3
  * or any future license endorsed by Mnemosyne LLC.
  *
- * $Id$
  */
 
 #include <assert.h>
@@ -470,7 +469,7 @@ tr_metainfoParseImpl (const tr_session  * session,
     {
       size_t len;
       char * bstr = tr_variantToStr (infoDict, TR_VARIANT_FMT_BENC, &len);
-      tr_sha1 (inf->hash, bstr, len, NULL);
+      tr_sha1 (inf->hash, bstr, (int) len, NULL);
       tr_sha1_to_hex (inf->hashString, inf->hash);
 
       if (infoDictLength != NULL)
